scribe2fhir/python/requirements.txt
# Python FHIR SDK Requirements

# Core dependencies
fhir-resources>=7.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Development dependencies (install with: pip install -e .[dev])
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# isort>=5.0.0
# mypy>=1.0.0
